+++ src/main/java/org/example/service/OldDescriptionService.java
@@ -1,7 +1,10 @@
 package org.example.service;
 
+import org.example.model.Description;
 import org.example.model.OldDescription;
 import org.example.repository.OldDescriptionRepository;
+import org.example.repository.ProductionOrderRepository;
+import org.example.repository.RewindOrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,13 +15,23 @@ import java.util.Optional;
 public class OldDescriptionService {
 
     private final OldDescriptionRepository oldDescriptionRepository;
+    private final ProductionOrderRepository productionOrderRepository;
+    private final RewindOrderRepository rewindOrderRepository;
 
     @Autowired
-    public OldDescriptionService(OldDescriptionRepository oldDescriptionRepository) {
+    public OldDescriptionService(OldDescriptionRepository oldDescriptionRepository,
+                                 ProductionOrderRepository productionOrderRepository,
+                                 RewindOrderRepository rewindOrderRepository) {
         this.oldDescriptionRepository = oldDescriptionRepository;
+        this.productionOrderRepository = productionOrderRepository;
+        this.rewindOrderRepository = rewindOrderRepository;
     }
 
     public OldDescription saveOldDescription(OldDescription oldDescription) {
+        Optional<OldDescription> existingOldDescription = oldDescriptionRepository.findByName(oldDescription.getName());
+        if (existingOldDescription.isPresent() && !existingOldDescription.get().getId().equals(oldDescription.getId())) {
+            throw new IllegalArgumentException("Old Name already exists");
+        }
         return oldDescriptionRepository.save(oldDescription);
     }
 
@@ -31,6 +44,9 @@ public class OldDescriptionService {
     }
 
     public void deleteOldDescription(OldDescription oldDescription) {
+        if (productionOrderRepository.existsByOldDescription(oldDescription) || rewindOrderRepository.existsByOldDescription(oldDescription)) {
+            throw new IllegalStateException("Cannot delete old description with existing orders");
+        }
         oldDescriptionRepository.delete(oldDescription);
     }
 }
